requirements.txt
openai
scikit-learn
streamlit
pandas
pydantic
python-docx
numpy